+++ examples/simple-example/web-server/app.js
@@ -1,7 +1,6 @@
 var express = require('express');
 var app = express();
 
-app.configure(function(){
   app.use(express.methodOverride());
   app.use(express.bodyParser());
   app.use(app.router);
@@ -9,18 +8,16 @@ app.configure(function(){
   app.set('views', __dirname + '/public');
   app.set('view options', {layout: false});
   app.set('basepath',__dirname + '/public');
-});
-
-app.configure('development', function(){
+let env = app.get('env');
+if(env=="development"){
   app.use(express.static(__dirname + '/public'));
   app.use(express.errorHandler({ dumpExceptions: true, showStack: true }));
-});
-
-app.configure('production', function(){
+}
+if(env == "production"){
   var oneYear = 31557600000;
   app.use(express.static(__dirname + '/public', { maxAge: oneYear }));
   app.use(express.errorHandler());
-});
+}
 
 console.log("Web server has started.\nPlease log on http://127.0.0.1:3001/index.html");