+++ examples/websocket-chat-ts-run/game-server/config/clientProtos.ts
@@ -1,1 +1,8 @@
-module.exports = {};
+module.exports = {
+    "chat.chatHandler.send": {
+        "required string content": 1,
+        "required string target": 2,
+        "required string rid": 3,
+        "required string from": 4,
+    }
+};